financial-ledger/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
pydantic==2.6.1
requests==2.31.0
